Add sort menu if features enabled
Hide it by default because of bug on the reloading the list
