+++ RunningOrder/Story/Views/StoryList.swift
@@ -15,8 +15,11 @@ extension StoryList {
         let sprint: Sprint
 
         @ObservedObject var logic: Logic
+
         @EnvironmentObject var storyManager: StoryManager
         @EnvironmentObject var searchManager: SearchManager
+        @EnvironmentObject var appStateManager: AppStateManager
+
         @State private var selected: Story?
         @State private var toBeDeletedStory: Story?
 
@@ -28,18 +31,21 @@ extension StoryList {
                             .epicColor(Color(identifier: logic.epicColor(for: story))),
                         tag: story,
                         selection: $selected,
-                        label: { StoryRow(story: story) }
-                    )
-                    .contextMenu {
-                        Button(action: { toBeDeletedStory = story }) {
-                            Text("Delete Story")
+                        label: {
+                            StoryRow(story: story)
+                                .frame(minHeight: 50, maxHeight: 150)
                         }
-                    }
-                    .epicColor(Color(identifier: logic.epicColor(for: story)))
+                    )
 
                     Divider()
                         .opacity(story == selected ? 0 : 1)
                 }
+                .contextMenu {
+                    Button(action: { toBeDeletedStory = story }) {
+                        Text("Delete Story")
+                    }
+                }
+                .epicColor(Color(identifier: logic.epicColor(for: story)))
             }
             .navigationTitle(logic.navigationTitle)
             .frame(minWidth: 100, idealWidth: 300)
@@ -47,7 +53,9 @@ extension StoryList {
                 ToolbarItems.sidebarItem
 
                 ToolbarItemGroup(placement: ToolbarItemPlacement.cancellationAction) {
-                    SortMenu(selectedSort: .constant(SortMenu.Option(type: .name, isReversed: false)))
+                    if appStateManager.enabledFeatures.contains(.sorting) {
+                        SortMenu(selectedSort: $storyManager.sortOption)
+                    }
 
                     Button(action: logic.showAddStoryView) {
                         Image(systemName: "square.and.pencil")
@@ -76,65 +84,20 @@ extension StoryList {
     }
 }
 
-struct SortMenu: View {
-    struct Option {
-        enum OptionType: String, CaseIterable {
-            case epic, name
-
-            var title: LocalizedStringKey {
-                switch self {
-                case .epic:
-                    return "Epic"
-                case .name:
-                    return "Name"
-                }
-            }
-        }
-        var type: OptionType
-        var isReversed: Bool
-
-        func apply(lhs: Story, rhs: Story) -> Bool {
-            switch (self.type, isReversed) {
-            case (.epic, false):
-                return lhs.epic < rhs.epic
-            case (.epic, true):
-                return lhs.epic > rhs.epic
-            case (.name, false):
-                return lhs.name < rhs.name
-            case (.name, true):
-                return lhs.name > rhs.name
-            }
-        }
-    }
-
-    @Binding var selectedSort: Option
-
-    func newIsReversed(for optionType: Option.OptionType) -> Bool {
-        return optionType == selectedSort.type && !selectedSort.isReversed
-    }
-
-    var body: some View {
-        Menu {
-            ForEach(Option.OptionType.allCases, id: \.rawValue) { optionType in
-                Button {
-                    selectedSort = Option(type: optionType, isReversed: newIsReversed(for: optionType))
-                } label: {
-                    Label(optionType.title, systemImage: optionType == selectedSort.type ? "checkmark" : "")
-                }
-            }
-        } label: {
-            Image(systemName: "line.horizontal.3.decrease.circle")
-        }
-    }
-}
-
 struct StoryList: View {
     let sprint: Sprint
     @EnvironmentObject var storyManager: StoryManager
     @EnvironmentObject var searchManager: SearchManager
 
     var body: some View {
-        InternalView(sprint: sprint, logic: Logic(storyManager: storyManager, searchManager: searchManager, sprint: sprint))
+        InternalView(
+            sprint: sprint,
+            logic: Logic(
+                storyManager: storyManager,
+                searchManager: searchManager,
+                sprint: sprint
+            )
+        )
     }
 }
